refactor: stock 라우터 단순화
