+++ backend/src/v2/routes.ts
@@ -4,7 +4,7 @@ import { initServer } from '@ts-rest/express';
 
 import { reviews } from './reviews/impl';
 import { histories } from './histories/impl';
-import { stock } from './stock/impl';
+import { stock } from './stock/mod.ts';
 
 const s = initServer();
 export default s.router(contract, {
