--- backend/src/v2/stock/controller/impl.ts
@@ -1,9 +1,0 @@
-import { mkGetStock, mkPatchStock } from './controller';
-import {
-  StockService,
-} from '../service';
-
-export const implStockController = (service: StockService) => ({
-  getStock: mkGetStock(service),
-  patchStock: mkPatchStock(service),
-});
