--- backend/src/v2/stock/controller/index.ts
@@ -1,1 +1,0 @@
-export * from './controller';
