--- backend/src/v2/stock/service/impl.ts
@@ -1,12 +1,0 @@
-import { Repository } from 'typeorm';
-import { VStock, Book } from '~/entity/entities';
-
-import { mkSearchStock, mkUpdateStock } from './service';
-
-export const implStockService = (repos: {
-	stockRepo: Repository<VStock>;
-  bookRepo: Repository<Book>;
-}) => ({
-  searchStock: mkSearchStock(repos),
-  updateStock: mkUpdateStock(repos),
-});
